Support armv8.(TODO: fix clock)

+++ Ubiquitous/XiZi_AIoT/hardkernel/clock/arm/armv8-a/cortex-a72/ok1028a-c/include/generic_timer.h
@@ -15,26 +15,26 @@
 #include <stdint.h>
 
 // armv8 generic timer
-static inline uint64_t r_cntv_ctl_el0()
+static inline uint32_t r_cntv_ctl_el0()
 {
-    uint64_t x;
+    uint32_t x;
     __asm__ volatile("mrs %0, cntv_ctl_el0" : "=r"(x));
     return x;
 }
 
-static inline void w_cntv_ctl_el0(uint64_t x)
+static inline void w_cntv_ctl_el0(uint32_t x)
 {
     __asm__ volatile("msr cntv_ctl_el0, %0" : : "r"(x));
 }
 
-static inline uint64_t r_cntv_tval_el0()
+static inline uint32_t r_cntv_tval_el0()
 {
-    uint64_t x;
+    uint32_t x;
     __asm__ volatile("mrs %0, cntv_tval_el0" : "=r"(x));
     return x;
 }
 
-static inline void w_cntv_tval_el0(uint64_t x)
+static inline void w_cntv_tval_el0(uint32_t x)
 {
     __asm__ volatile("msr cntv_tval_el0, %0" : : "r"(x));
 }
@@ -46,9 +46,9 @@ static inline uint64_t r_cntvct_el0()
     return x;
 }
 
-static inline uint64_t r_cntfrq_el0()
+static inline uint32_t r_cntfrq_el0()
 {
-    uint64_t x;
+    uint32_t x;
     __asm__ volatile("mrs %0, cntfrq_el0" : "=r"(x));
     return x;
 }
